+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Exclude Stratis from table
 
-## Fixed
+### Fixed
 
 - Missing friendly map names and links