+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 
-## [UNRELEASED] - TBC
+## [0.7.0] - 20205-04-17
 
 ### Added
 
@@ -100,7 +100,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 Initial release
 
 
-[UNRELEASED]: https://github.com/recreational-projects/A3-Antistasi-Ultimate-analysis/compare/v0.6.0...HEAD
+[0.7.0]: https://github.com/recreational-projects/A3-Antistasi-Ultimate-analysis/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/recreational-projects/A3-Antistasi-Ultimate-analysis/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/recreational-projects/A3-Antistasi-Ultimate-analysis/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/recreational-projects/A3-Antistasi-Ultimate-analysis/compare/v0.3.1...v0.4.0